requirements.txt
rasterio
tensorflow[and-cuda]
numpy
pandas
matplotlib
scikit-learn
streamlit
Pillow
opencv
scipy
